[core] Merge FileBasedTileSource and TileSource
