--- src/mbgl/tile/tile_source.cpp
@@ -1,7 +1,0 @@
-#include <mbgl/tile/tile_source.hpp>
-
-#include <cassert>
-
-namespace mbgl {
-
-} // namespace mbgl
