+++ src/mbgl/tile/tile_source.hpp
@@ -1,9 +1,20 @@
 #pragma once
 
 #include <mbgl/util/noncopyable.hpp>
+#include <mbgl/storage/resource.hpp>
 
 namespace mbgl {
 
+class FileSource;
+class AsyncRequest;
+class Response;
+class Tileset;
+
+namespace style {
+class UpdateParameters;
+}
+
+template <typename T>
 class TileSource : private util::noncopyable {
 public:
     // TileSources can have two states: optional or required.
@@ -16,20 +27,10 @@ public:
         Required = true,
     };
 
-protected:
-    TileSource() : necessity(Necessity::Optional) {
-    }
-
-public:
-    virtual ~TileSource() = default;
-
-    bool isOptional() const {
-        return necessity == Necessity::Optional;
-    }
-
-    bool isRequired() const {
-        return necessity == Necessity::Required;
-    }
+    TileSource(T&,
+               const OverscaledTileID&,
+               const style::UpdateParameters&,
+               const Tileset&);
 
     void setNecessity(Necessity newNecessity) {
         if (newNecessity != necessity) {
@@ -42,19 +43,34 @@ public:
         }
     }
 
-protected:
+private:
+    bool isOptional() const {
+        return necessity == Necessity::Optional;
+    }
+
+    bool isRequired() const {
+        return necessity == Necessity::Required;
+    }
+
     // called when the tile is one of the ideal tiles that we want to show definitely. the tile source
     // should try to make every effort (e.g. fetch from internet, or revalidate existing resources).
-    virtual void makeRequired() {}
+    void makeRequired();
 
     // called when the zoom level no longer corresponds to the displayed one, but
     // we're still interested in retaining the tile, e.g. for backfill.
     // subclassed TileSources should cancel actions they are taking to provide
     // an up-to-date version or load new data
-    virtual void makeOptional() {}
+    void makeOptional();
 
-protected:
+    void loadOptional();
+    void loadedData(const Response&);
+    void loadRequired();
+
+    T& tileData;
     Necessity necessity;
+    Resource resource;
+    FileSource& fileSource;
+    std::unique_ptr<AsyncRequest> request;
 };
 
 } // namespace mbgl
