+++ src/mbgl/tile/vector_tile_data.cpp
@@ -1,7 +1,6 @@
 #include <mbgl/tile/vector_tile_data.hpp>
-#include <mbgl/tile/file_based_tile_source.hpp>
-#include <mbgl/tile/file_based_tile_source_impl.hpp>
 #include <mbgl/tile/tile_source.hpp>
+#include <mbgl/tile/tile_source_impl.hpp>
 #include <mbgl/tile/vector_tile.hpp>
 #include <mbgl/style/update_parameters.hpp>
 #include <mbgl/util/async_request.hpp>
@@ -13,13 +12,13 @@ VectorTileData::VectorTileData(const OverscaledTileID& id_,
                                const style::UpdateParameters& parameters,
                                const Tileset& tileset)
     : GeometryTileData(id_, sourceID, parameters.style, parameters.mode),
-      tileSource(std::make_unique<FileBasedTileSource<VectorTileData>>(*this, id_, parameters, tileset)) {
+      tileSource(*this, id_, parameters, tileset) {
 }
 
 VectorTileData::~VectorTileData() = default;
 
 void VectorTileData::setNecessity(Necessity necessity) {
-    tileSource->setNecessity(static_cast<TileSource::Necessity>(necessity));
+    tileSource.setNecessity(static_cast<TileSource<VectorTileData>::Necessity>(necessity));
 }
 
 std::unique_ptr<GeometryTile> VectorTileData::parseData(std::shared_ptr<const std::string> data) {
